# Bibliothek/swagger_doc.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Library API

host: localhost:3000
basePath: /

schemes:
  - http

paths:
  /:
    get:
      summary: Welcome endpoint
      responses:
        200:
          description: OK
          schema:
            type: string

  /books:
    get:
      summary: Get all books
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Book"
    post:
      summary: Create a new book
      parameters:
        - name: newBook
          in: body
          description: The book to be created
          required: true
          schema:
            $ref: "#/definitions/Book"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Book"
  
  /books/{isbn}:
    get:
      summary: Get a book by ISBN
      parameters:
        - name: isbn
          in: path
          description: The ISBN of the book
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Book"
        404:
          description: Book not found

    put:
      summary: Update a book by ISBN
      parameters:
        - name: isbn
          in: path
          description: The ISBN of the book
          required: true
          type: integer
        - name: updatedBook
          in: body
          description: The updated book
          required: true
          schema:
            $ref: "#/definitions/Book"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Book"
        404:
          description: Book not found

    delete:
      summary: Delete a book by ISBN
      parameters:
        - name: isbn
          in: path
          description: The ISBN of the book
          required: true
          type: integer
      responses:
        200:
          description: Book successfully deleted
        404:
          description: Book not found

  /lends:
    get:
      summary: Get all lends
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Lend"
    post:
      summary: Create a new lend
      parameters:
        - name: newLend
          in: body
          description: The lend to be created
          required: true
          schema:
            $ref: "#/definitions/Lend"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Lend"

  /lends/{id}:
    get:
      summary: Get a lend by ID
      parameters:
        - name: id
          in: path
          description: The ID of the lend
          required: true
          type: integer
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Lend"
        404:
          description: Lend not found

    patch:
      summary: Update a lend by ID
      parameters:
        - name: id
          in: path
          description: The ID of the lend
          required: true
          type: integer
        - name: updatedLend
          in: body
          description: The updated lend
          required: true
          schema:
            $ref: "#/definitions/Lend"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Lend"
        404:
          description: Lend not found

definitions:
  Book:
    type: object
    properties:
      isbn:
        type: integer
      title:
        type: string
      year:
        type: string
      author:
        type: string

  Lend:
    type: object
    properties:
      id:
        type: integer
      customer_id:
        type: integer
      isbn:
        type: integer
      borrowed_at:
        type: string
      returned_at:
        type: string